Removing utils package and moving Redact to the debug package
