+++ debug/debug_test.go
@@ -3,6 +3,8 @@ package debug
 import (
 	"bytes"
 	"testing"
+
+	"github.com/stretchr/testify/assert"
 )
 
 func TestVerboseEnabled(t *testing.T) {
@@ -25,4 +27,11 @@ func TestVerboseDisabled(t *testing.T) {
 	if b.String() != "" {
 		t.Error("expected '' got", b.String())
 	}
+}
+
+func TestRedact(t *testing.T) {
+	fakeToken := "1a11111aaaa111aa1a11111a11111aa1"
+	expected := "1a11*************************aa1"
+
+	assert.Equal(t, expected, Redact(fakeToken))
 }
